+++ src/scripts/create_multi_wiki_qa.py
@@ -43,6 +43,7 @@ LANGUAGES = [
     "pt-pt",
     "lv",
     "lt",
+    "sk",
 ]