+++ src/scripts/create_winogrande.py
@@ -43,6 +43,7 @@ LANGUAGES = [
     "no",
     "pl",
     "pt",
+    "sk",
     "sv",
 ]