+++ src/scripts/load_ud_pos.py
@@ -380,6 +380,42 @@ def load_ltdt_pos() -> dict[str, pd.DataFrame]:
     return load_ud_pos(train_url=train_url, val_url=val_url, test_url=test_url)
 
 
+def load_csdt_pos() -> dict[str, pd.DataFrame]:
+    """Load the part-of-speech part of the Czech Dependency Treebank.
+
+    Returns:
+        The dataframes, stored in the keys `train`, `val` and `test`.
+    """
+    # Define download URLs
+    base_url = (
+        "https://raw.githubusercontent.com/UniversalDependencies/UD_Czech-CAC/refs/"
+        "heads/master/cs_cac-ud-{}.conllu"
+    )
+    train_url = base_url.format("train")
+    val_url = base_url.format("dev")
+    test_url = base_url.format("test")
+
+    return load_ud_pos(train_url=train_url, val_url=val_url, test_url=test_url)
+
+
+def load_skdt_pos() -> dict[str, pd.DataFrame]:
+    """Load the part-of-speech part of the Slovak Dependency Treebank.
+
+    Returns:
+        The dataframes, stored in the keys `train`, `val`, and `test`.
+    """
+    # Define download URLs
+    base_url = (
+        "https://raw.githubusercontent.com/UniversalDependencies/UD_Slovak-SNK/refs/heads/master/"
+        "sk_snk-ud-{}.conllu"
+    )
+    train_url = base_url.format("train")
+    val_url = base_url.format("dev")
+    test_url = base_url.format("test")
+
+    return load_ud_pos(train_url=train_url, val_url=val_url, test_url=test_url)
+
+
 def _load_file_or_url(url_or_path: str) -> list[str]:
     """Load a file from a URL or local path.
 
@@ -506,24 +542,6 @@ def _load_split(
 
     # Convert the records to a dataframe
     return pd.DataFrame.from_records(records)
-
-
-def load_csdt_pos() -> dict[str, pd.DataFrame]:
-    """Load the part-of-speech part of the Czech Dependency Treebank.
-
-    Returns:
-        The dataframes, stored in the keys `train`, `val` and `test`.
-    """
-    # Define download URLs
-    base_url = (
-        "https://raw.githubusercontent.com/UniversalDependencies/UD_Czech-CAC/refs/"
-        "heads/master/cs_cac-ud-{}.conllu"
-    )
-    train_url = base_url.format("train")
-    val_url = base_url.format("dev")
-    test_url = base_url.format("test")
-
-    return load_ud_pos(train_url=train_url, val_url=val_url, test_url=test_url)
 
 
 def load_ud_pos(