+++ CHANGELOG.md
@@ -56,6 +56,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added support for Slovak 🇸🇰! This includes the sentiment classification dataset
+  CSFD-sentiment-sk, the linguistic acceptability dataset ScaLA-sk, the named entity
+  recognition dataset UNER-sk, the reading comprehension dataset MultiWikiQA-sk, the
+  multiple-choice classification dataset MMLU-sk, and the common-sense reasoning dataset
+  Winogrande-sk.
 - Added support for Czech 🇨🇿! This includes the sentiment classification dataset
   CSFD-sentiment, the linguistic acceptability dataset ScaLA-cs, the
   linguistic acceptability dataset CS-GEC,